+++ src/main/java/org/nrg/containers/model/command/auto/Command.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.auto.value.AutoValue;
 import com.google.common.base.Function;
 import com.google.common.collect.ImmutableList;
@@ -19,11 +21,8 @@ import org.nrg.containers.model.configuration.CommandConfiguration.CommandOutput
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.io.IOException;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -54,7 +53,7 @@ public abstract class Command {
     @Nullable @JsonProperty("limit-cpu") public abstract Double limitCpu();
 
     @JsonIgnore private static Pattern regCharPattern = Pattern.compile("[^A-Za-z0-9_-]");
-
+    @JsonIgnore private static ObjectMapper mapper = new ObjectMapper();
 
     @JsonCreator
     static Command create(@JsonProperty("id") final long id,
@@ -301,6 +300,7 @@ public abstract class Command {
         final String knownMounts = StringUtils.join(mountNames, ", ");
 
         final Set<String> inputNames = Sets.newHashSet();
+        final Map<String, CommandInput> commandInputs = Maps.newHashMap();
         for (final CommandInput input : inputs()) {
             final List<String> inputErrors = Lists.newArrayList();
             inputErrors.addAll(Lists.transform(input.validate(), addCommandNameToError));
@@ -309,6 +309,7 @@ public abstract class Command {
                 errors.add(commandName + "input name \"" + input.name() + "\" is not unique.");
             } else {
                 inputNames.add(input.name());
+                commandInputs.put(input.name(), input);
             }
 
             if (!inputErrors.isEmpty()) {
@@ -328,7 +329,8 @@ public abstract class Command {
             }
 
             if (!mountNames.contains(output.mount())) {
-                errors.add(commandName + "output \"" + output.name() + "\" references unknown mount \"" + output.mount() + "\". Known mounts: " + knownMounts);
+                errors.add(commandName + "output \"" + output.name() + "\" references unknown mount \"" +
+                        output.mount() + "\". Known mounts: " + knownMounts);
             }
 
             if (!outputErrors.isEmpty()) {
@@ -366,6 +368,17 @@ public abstract class Command {
                     wrapperInputNames.add(external.name());
                 }
 
+                CommandInput provFor = commandInputs.get(external.providesValueForCommandInput());
+                if (provFor != null) {
+                    if (provFor.isSelect()) {
+                        errors.add(wrapperName + "external input name \"" + external.name() + "\" provides value for " +
+                                "command input \"" + provFor.name() + "\", which is marked as a select type.");
+                    }
+                    if (!provFor.selectValues().isEmpty()) {
+                        errors.add(wrapperName + "external input name \"" + external.name() + "\" provides value for " +
+                                "command input \"" + provFor.name() + "\", which has select values.");
+                    }
+                }
 
                 if (!inputErrors.isEmpty()) {
                     errors.addAll(inputErrors);
@@ -388,7 +401,21 @@ public abstract class Command {
                     errors.add(wrapperName + "derived input \"" + derived.name() + "\" is derived from itself.");
                 }
                 if (!wrapperInputNames.contains(derived.derivedFromWrapperInput())) {
-                    errors.add(wrapperName + "derived input \"" + derived.name() + "\" is derived from an unknown XNAT input \"" + derived.derivedFromWrapperInput() + "\". Known inputs: " + StringUtils.join(wrapperInputNames, ", "));
+                    errors.add(wrapperName + "derived input \"" + derived.name() +
+                            "\" is derived from an unknown XNAT input \"" + derived.derivedFromWrapperInput() +
+                            "\". Known inputs: " + StringUtils.join(wrapperInputNames, ", "));
+                }
+
+                CommandInput provFor = commandInputs.get(derived.providesValueForCommandInput());
+                if (provFor != null) {
+                    if (provFor.isSelect()) {
+                        errors.add(wrapperName + "derived input name \"" + derived.name() + "\" provides value for " +
+                                "command input \"" + provFor.name() + "\", which is marked as a select type.");
+                    }
+                    if (!provFor.selectValues().isEmpty()) {
+                        errors.add(wrapperName + "derived input name \"" + derived.name() + "\" provides value for " +
+                                "command input \"" + provFor.name() + "\", which has select values.");
+                    }
                 }
 
                 if (!inputErrors.isEmpty()) {
@@ -406,13 +433,15 @@ public abstract class Command {
                 outputErrors.addAll(Lists.transform(output.validate(), addWrapperNameToError));
 
                 if (!outputNames.contains(output.commandOutputName())) {
-                    errors.add(wrapperName + "output handler refers to unknown command output \"" + output.commandOutputName() + "\". Known outputs: " + knownOutputs + ".");
+                    errors.add(wrapperName + "output handler refers to unknown command output \"" +
+                            output.commandOutputName() + "\". Known outputs: " + knownOutputs + ".");
                 } else {
                     handledOutputs.add(output.commandOutputName());
                 }
 
                 if (!(wrapperInputNames.contains(output.targetName()) || wrapperOutputNames.contains(output.targetName()))) {
-                    errors.add(wrapperName + "output handler does not refer to a known wrapper input or output. \"as-a-child-of\": \"" + output.targetName() + "\"." +
+                    errors.add(wrapperName + "output handler does not refer to a known wrapper input or output. " +
+                            "\"as-a-child-of\": \"" + output.targetName() + "\"." +
                             "\nKnown inputs: " + knownWrapperInputs + "." +
                             "\nKnown outputs (so far): " + StringUtils.join(wrapperOutputNames, ", ") + ".");
                 }
@@ -439,7 +468,8 @@ public abstract class Command {
                 // We know at least one output is not handled. Now find out which.
                 for (final String commandOutput : outputNames) {
                     if (!handledOutputs.contains(commandOutput)) {
-                        errors.add(wrapperName + "command output \"" + commandOutput + "\" is not handled by any output handler.");
+                        errors.add(wrapperName + "command output \"" + commandOutput +
+                                "\" is not handled by any output handler.");
                     }
                 }
             }
@@ -612,6 +642,7 @@ public abstract class Command {
         @Nullable @JsonProperty("command-line-separator") public abstract String commandLineSeparator();
         @Nullable @JsonProperty("true-value") public abstract String trueValue();
         @Nullable @JsonProperty("false-value") public abstract String falseValue();
+        @JsonProperty("select-values") public abstract ImmutableList<String> selectValues();
         @Nullable @JsonProperty("multiple-delimiter") public abstract String multipleDelimiter();
 
         @JsonCreator
@@ -628,6 +659,7 @@ public abstract class Command {
                                    @JsonProperty("true-value") final String trueValue,
                                    @JsonProperty("false-value") final String falseValue,
                                    @JsonProperty("sensitive") final Boolean sensitive,
+                                   @JsonProperty("select-values") final List<String> selectValues,
                                    @JsonProperty("multiple-delimiter") final String multipleDelimiter) {
             return builder()
                     .name(name)
@@ -643,6 +675,7 @@ public abstract class Command {
                     .trueValue(trueValue)
                     .falseValue(falseValue)
                     .sensitive(sensitive)
+                    .selectValues(selectValues == null ? Collections.emptyList() : selectValues)
                     .multipleDelimiter(multipleDelimiter)
                     .build();
         }
@@ -667,6 +700,7 @@ public abstract class Command {
                     .trueValue(commandInputEntity.getTrueValue())
                     .falseValue(commandInputEntity.getFalseValue())
                     .sensitive(commandInputEntity.getSensitive())
+                    .selectValues(commandInputEntity.getSelectValues())
                     .multipleDelimiter(md == null ? null : md.getName())
                     .build();
         }
@@ -676,7 +710,8 @@ public abstract class Command {
                     .id(0L)
                     .name("")
                     .type(CommandInputEntity.DEFAULT_TYPE.getName())
-                    .required(false);
+                    .required(false)
+                    .selectValues(Collections.emptyList());
         }
 
         public CommandInput applyConfiguration(final CommandInputConfiguration commandInputConfiguration) {
@@ -693,10 +728,23 @@ public abstract class Command {
                     .trueValue(this.trueValue())
                     .falseValue(this.falseValue())
                     .sensitive(this.sensitive())
+                    .selectValues(this.selectValues())
                     .multipleDelimiter(this.multipleDelimiter())
                     .defaultValue(commandInputConfiguration.defaultValue())
                     .matcher(commandInputConfiguration.matcher())
                     .build();
+        }
+
+        @JsonIgnore
+        public boolean isSelect() {
+            String type = type();
+            return type.equals(CommandInputEntity.Type.MULTISELECT.getName()) ||
+                    type.equals(CommandInputEntity.Type.SELECT.getName());
+        }
+
+        @JsonIgnore
+        public boolean isMultiSelect() {
+            return type().equals(CommandInputEntity.Type.MULTISELECT.getName());
         }
 
         public abstract Builder toBuilder();
@@ -718,6 +766,33 @@ public abstract class Command {
                 errors.add("Invalid multiple-delimiter \"" + md + "\", choose from: " +
                         StringUtils.join(names, ", "));
             }
+            List<String> selectValues = selectValues();
+            if (isSelect() && selectValues.isEmpty()) {
+                errors.add("Command input \"" +  name()  + "\" is designated as type " + type() + " but doesn't list " +
+                        "select-values. Note that command inputs with values provided by xnat inputs shouldn't be " +
+                        "designated as select (they'll automatically render as a select if their xnat input resolves " +
+                        "to more than one value.");
+            } else if (!isSelect() && !selectValues.isEmpty()) {
+                errors.add("Command input \"" +  name()  + "\" has select-values set, but is not a select type.");
+            }
+            String defaultVal = defaultValue();
+            List<String> defaultValList = null;
+            if (isMultiSelect()) {
+                try {
+                    defaultValList = mapper.readValue(defaultVal, new TypeReference<List<String>>() {});
+                } catch (IOException e) {
+                    // Not a list, treat defaultVal as string
+                }
+            }
+            if (defaultValList != null) {
+                if (!selectValues.isEmpty() && !selectValues.containsAll(defaultValList)) {
+                    errors.add("Command input \"" + name() + "\" one or more default values in \"" + defaultValList +
+                            "\" is not in the list of select-values \"" + selectValues + "\"");
+                }
+            } else if (defaultVal != null && !selectValues.isEmpty() && !selectValues.contains(defaultVal)) {
+                errors.add("Command input \"" +  name()  + "\" default value \"" + defaultVal + "\" is not one of the " +
+                        "select-values \"" + selectValues + "\"");
+            }
             return errors;
         }
 
@@ -737,7 +812,8 @@ public abstract class Command {
             public abstract Builder trueValue(final String trueValue);
             public abstract Builder falseValue(final String falseValue);
             public abstract Builder sensitive(Boolean sensitive);
-            public abstract Builder multipleDelimiter(String multipleDelimiter);
+            public abstract Builder selectValues(final List<String> selectValues);
+            public abstract Builder multipleDelimiter(final String multipleDelimiter);
 
             public abstract CommandInput build();
         }