+++ src/main/java/org/nrg/containers/model/command/auto/ResolvedCommand.java
@@ -134,11 +134,20 @@ public abstract class ResolvedCommand {
         commandInputValues = commandInputValuesBuilder.build();
     }
 
-    @Nonnull
     @JsonIgnore
     public static Command.CommandInput collectCommandInputChildrenOfMultipleDerivedInput(final Command.CommandWrapperInput wrapperInput,
                                                                                          final List<ResolvedInputTreeNode.ResolvedInputTreeValueAndChildren> derivedInputValuesAndChildren,
                                                                                          final List<String> commandInputChildrenValues)
+            throws CommandResolutionException {
+        return collectCommandInputChildrenOfMultipleDerivedInput(wrapperInput, derivedInputValuesAndChildren,
+                commandInputChildrenValues, true);
+    }
+
+    @JsonIgnore
+    public static Command.CommandInput collectCommandInputChildrenOfMultipleDerivedInput(final Command.CommandWrapperInput wrapperInput,
+                                                                                         final List<ResolvedInputTreeNode.ResolvedInputTreeValueAndChildren> derivedInputValuesAndChildren,
+                                                                                         final List<String> commandInputChildrenValues,
+                                                                                         final boolean throwExceptionIfNullAndReqd)
             throws CommandResolutionException {
 
         String commandInputName = wrapperInput.providesValueForCommandInput();
@@ -169,7 +178,12 @@ public abstract class ResolvedCommand {
             }
         }
 
-        if (ci == null) {
+        if (ci == null && throwExceptionIfNullAndReqd && wrapperInput.required()) {
+            // ci == null will occur if the derived input has no resolved values e.g. nothing matches the matcher.
+            // For the UI (a.k.a., when this is run for preresolution), we want to make the user aware of the issue,
+            // which is accomplished with a warning from commandUiLauncher.js. So, we don't want to throw an exception.
+            // Otherwise (during command resolution), we probably want to throw an exception, but if the parent input
+            // isn't required, then perhaps it's okay...
             throw new CommandResolutionException(wrapperInput.name() + " must have precisely one command " +
                     "input child element");
         }
@@ -211,7 +225,9 @@ public abstract class ResolvedCommand {
                         ResolvedInputTreeNode.ResolvedInputTreeValueAndChildren.create(ResolvedInputValue.builder()
                                 .value(String.join(" ", commandInputChildrenValues))
                                 .build());
-                flatTree.add(ResolvedInputTreeNode.create(ci, Collections.singletonList(valueAndChildren)));
+                if (ci != null) {
+                    flatTree.add(ResolvedInputTreeNode.create(ci, Collections.singletonList(valueAndChildren)));
+                }
             } catch (CommandResolutionException e) {
                 // Ignore for purposes of flattening tree
             }