+++ src/main/java/org/nrg/containers/model/command/entity/CommandInputEntity.java
@@ -5,19 +5,9 @@ import org.nrg.containers.model.command.auto.Command;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Transient;
+import javax.persistence.*;
 import java.io.Serializable;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Entity
@@ -39,6 +29,7 @@ public class CommandInputEntity implements Serializable {
     private String trueValue;
     private String falseValue;
     private Boolean sensitive;
+    private List<String> selectValues = Collections.emptyList();
     private MultipleDelimiter multipleDelimiter;
 
     public static CommandInputEntity fromPojo(final Command.CommandInput commandInput) {
@@ -62,6 +53,7 @@ public class CommandInputEntity implements Serializable {
         this.setTrueValue(commandInput.trueValue());
         this.setFalseValue(commandInput.falseValue());
         this.setSensitive(commandInput.sensitive());
+        this.setSelectValues(commandInput.selectValues());
         this.setMultipleDelimiterByName(commandInput.multipleDelimiter());
 
         switch (commandInput.type()) {
@@ -73,6 +65,12 @@ public class CommandInputEntity implements Serializable {
                 break;
             case "number":
                 this.setType(Type.NUMBER);
+                break;
+            case "select-one":
+                this.setType(Type.SELECT);
+                break;
+            case "select-many":
+                this.setType(Type.MULTISELECT);
                 break;
             default:
                 this.setType(DEFAULT_TYPE);
@@ -229,6 +227,15 @@ public class CommandInputEntity implements Serializable {
         }
     }
 
+    @ElementCollection
+    public List<String> getSelectValues() {
+        return selectValues;
+    }
+
+    public void setSelectValues(List<String> selectValues) {
+        this.selectValues = selectValues;
+    }
+
     @Override
     public boolean equals(final Object o) {
         if (this == o) return true;
@@ -260,6 +267,7 @@ public class CommandInputEntity implements Serializable {
                 .add("trueValue", trueValue)
                 .add("falseValue", falseValue)
                 .add("sensitive", sensitive)
+                .add("selectValues", selectValues)
                 .add("multipleDelimiter", multipleDelimiter)
                 .toString();
     }
@@ -267,7 +275,9 @@ public class CommandInputEntity implements Serializable {
     public enum Type {
         STRING("string"),
         BOOLEAN("boolean"),
-        NUMBER("number");
+        NUMBER("number"),
+        SELECT("select-one"),
+        MULTISELECT("select-many");
 
         public final String name;